# store/cache: speed up + conserve memory by using map clearing idiom in Reset